fixup! drivers/at: make OK and ERROR replies configurable

+++ drivers/at/at.c
@@ -141,15 +141,15 @@ ssize_t at_send_cmd_get_lines(at_dev_t *dev, const char *command,
         }
         else if (res > 0) {
             bytes_left -= res;
-            ssize_t len_ok = sizeof(AT_RECV_OK) - 1;
-            ssize_t len_error = sizeof(AT_RECV_ERROR) - 1;
-            if ((res == (len_ok + keep_eol)) &&
+            size_t len_ok = sizeof(AT_RECV_OK) - 1;
+            size_t len_error = sizeof(AT_RECV_ERROR) - 1;
+            if (((size_t )res == (len_ok + keep_eol)) &&
                 (len_ok != 0) &&
                 (strncmp(pos, AT_RECV_OK, len_ok) == 0)) {
                 res = len - bytes_left;
                 break;
             }
-            else if ((res == (len_error + keep_eol)) &&
+            else if (((size_t )res == (len_error + keep_eol)) &&
                      (len_error != 0) &&
                      (strncmp(pos, AT_RECV_ERROR, len_error) == 0)) {
                 return -1;
